raise execpetion on save! if valid? fails
