+++ test/document_test.rb
@@ -1,41 +1,5 @@
 require 'test_helper'
-
-class BaseDolly < Dolly::Document; end
-
-class BarFoo < BaseDolly
-  property :a, :b, :c, :d, :e, :f, :g, :h, :i, :j, :k, :l, :m, :n, :persist
-end
-
-class FooBar < BaseDolly
-  property :foo, :bar
-  property :with_default, default: 1
-  property :boolean, class_name: TrueClass, default: true
-  property :date, class_name: Date
-  property :time, class_name: Time
-  property :datetime, class_name: DateTime
-  property :is_nil, class_name: NilClass, default: nil
-
-  timestamps!
-end
-
-class Baz < Dolly::Document; end
-
-class FooBaz < Dolly::Document
-  property :foo, class_name: Hash, default: {}
-
-  def add_to_foo key, value
-    foo[key] ||= value
-    save!
-  end
-end
-
-class WithTime < Dolly::Document
-  property :created_at
-end
-
-class TestFoo < Dolly::Document
-  property :default_test_property, class_name: String, default: 'FOO'
-end
+require 'support/base_documents'
 
 class DocumentTest < ActiveSupport::TestCase
   DB_BASE_PATH = "http://localhost:5984/test".freeze
@@ -393,6 +357,13 @@ class DocumentTest < ActiveSupport::TestCase
     foo.bar = 'I belong to the foo, not the bar'
     bar = FooBar.new
     assert_not_equal foo.bar, bar.bar
+  end
+
+  test 'subclass raises DocumentInvalidError if valid? fails' do
+    foo = ObjectWithValid.new
+    assert_raise Dolly::DocumentInvalidError do
+      foo.save!
+    end
   end
 
   private
